refactor: globalization of types
